refactor: User package generate and files move
