+++ src/main/java/yagu/yagu/common/security/CustomOAuth2User.java
@@ -5,7 +5,7 @@ import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.oauth2.core.user.OAuth2User;
-import yagu.yagu.entity.User;
+import yagu.yagu.user.entity.User;
 
 import java.util.Collection;
 import java.util.Collections;
