+++ src/main/java/yagu/yagu/common/security/CustomOAuth2UserService.java
@@ -6,8 +6,8 @@ import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
 import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
-import yagu.yagu.entity.User;
-import yagu.yagu.service.AuthService;
+import yagu.yagu.user.entity.User;
+import yagu.yagu.user.service.AuthService;
 
 import java.util.Map;
