+++ src/main/java/yagu/yagu/common/security/CustomUserDetailsService.java
@@ -5,8 +5,8 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-import yagu.yagu.entity.User;
-import yagu.yagu.repository.UserRepository;
+import yagu.yagu.user.entity.User;
+import yagu.yagu.user.repository.UserRepository;
 
 import java.util.Collections;
