+++ src/main/java/yagu/yagu/common/security/SecurityConfig.java
@@ -12,7 +12,7 @@ import org.springframework.security.config.annotation.web.configurers.AbstractHt
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import yagu.yagu.service.AuthService;
+import yagu.yagu.user.service.AuthService;
 
 import java.nio.charset.StandardCharsets;
 import java.util.Map;
