--- src/main/java/yagu/yagu/controller/AuthController.java
@@ -1,4 +1,0 @@
-package yagu.yagu.controller;
-
-public class AuthController {
-}
