+++ src/main/java/yagu/yagu/user/controller/AuthController.java
@@ -1,0 +1,4 @@
+package yagu.yagu.user.controller;
+
+public class AuthController {
+}
