+++ src/main/java/yagu/yagu/user/entity/User.java
@@ -1,4 +1,4 @@
-package yagu.yagu.entity;
+package yagu.yagu.user.entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
